Moved configurations from the main API package into dedicated packages
